refactor: 캐스팅 페이지 네비게이션 컴포넌트로 분리
